<3TbsJegElskerDig!! 🌌🌍 [ graphics3d ] [ day 51 ] [ C++ ] [ ca2 ] SceneFoundry applications at ca2 Store <3TbsJegElskerDig!!
<3tbs, Mummi and bilbo!!

Thomas Borregaard Sørensen \infinity,-0.16091989,\infinity ONE-MAN ABSOLUTE <3!! I love you, by ???-0.02041977-???write my history please make me please create me for you for me for you for me Camilo Sasuke Thomas Borregaard Sørensen!!

Thomas 3 private commits on mid Dec2020!!

Thomas Online YouTube VODs contribution!!

Mummi orange-rice-flour cake on 20-Dec!!

Mummi : Thank you!! Keeping things tidy, keeping things clean, keep ca2 build clean, what's that error camilo... !!

Mummi : Thank you!! Release to the client, he is waiting!! Don't make him wait!! Attend well the client like he were a king!!

Mummi (tinytaura) watching and chatting contribution!!

bilbo sleeping and needing/requesting/crying for help care (for the right person (me), the cats wanna fight with him) contribution!!

sodapoppin and friends contribution!!

V0idsEmbrace contribution while assisting to code graphics3d (vulkan, opengl, directx11 and directx12)!!
and Cherno YT about layers ->
instead of beginframe begin_frame end_frame end_frame
more organized naming and ideas :  begin frame -> start layer end_layer  -> start_layer end_layer ... -> end_frame

lnvhj@youtube chatting contribution!!

iAssyrian chatting contribution!!

boflux (Spoofh, Benjamin Kuhl) chatting contribution!!

jusg_fpga (fpga_guru, vue_equalizer, just_fpga, Oliver Pohl) chatting contribution!!

cmgriffing streaming, raiding and raiders contribution!!

robotichead streaming and playing games with cmgriffing contribution!!

TimBeaudet (Friends: FletcherLabs, tsjost and Jabokoe) streaming contribution!!

Stumpen_nicklas_dk, sodapoppin and EduardoRFS streaming contribution!!

Roxkstar74 sleeping streaming contribution!!

kissloryshy chatting contribution!!

MrShoor (from Russia) streaming C++ coding contribution!!

blackjekko from Padova Italia through twitch C++/ca2 interest contribution!!

j_blow streaming contribution!!

K2TheFam (@Twitch) streaming and chatting contribution!!

boflux (Ben, Spoofh, from Germany) chatting contribution!!

parrot_rl chatting contribution (from New Jersey)!!

JPCdk streaming contribution!!

whyyyyyyysoserious streaming chess contribution!!

fpga_guru (vue_equalizer, Oliver from Deutsch)  C++/ca2 interest contribution!!

SovereignDev with Unreal streaming contribution!!

Ash_F0x and TimBeaudet streaming contribution!!

Myrkee (Valheim) streaming contribution!!

xmetrix and EinfachUwe42 streaming contribution!!

JessicaMak and marcobrunodev streaming contribution!!

alfredotigolo, mandrakenk and Okbatgames chatting contribution!!

jitspoe, Endesga and Fearitself streaming contribution!!

jmcmorris (Jason Morris, SiegeGames) streaming contribution!!

tomrandall streaming Ludum contribution!!

vue_equalizer (fpga_guru) chatting contribution!!

Thiagovgamg chatting contribution!!

Naysayer88 and friends contribution!!

lelandkwong streaming contribution!!

Goldbargames streaming contribution!!

Bytakos (bytakos) streaming contribution!!

Endesga streaming contribution!!

TravisVroman streaming, raid(2025-12-11 and raiders) and follow contribution!!

jitspoe and strager streaming contribution!!

Ash_F0x and JessicaMak streaming contribution!!

WTSRetro/SpiffyDane and Myrkee streaming contribution!!

Ninja and friends streaming contribution!!

erald_guri chatting contribution!!

lastmiles streaming farwest contribution!!

Gram streaming contribution!!

banikaze@twitch (fra Danmark) streaming games contribution!!

rw_grim streaming contribution!!

AdamCYounis streaming contribution!!

Dunno (P4ndaExpress) chatting and streaming contribution!!

Zorchenhimer streaming contribution!!

lasteveq4 C++ interest chat contriubtion!!

cecilphillip and clarkio @"Microsoft Developer" streaming contribution!!

oijtx streaming contribution!!

lowluds chatting, chilling, paying attention, active support and channel interest contribution!!

Gabiul (7Electron, shiroespark with Spring/Java) chatting contribution!!

FrankL81 talking about ca2 and C++ frameworks and possible virtualization of ca2 on platforms contribution!!

t104mo contribution!!

shegotmad (@twitch) learning C++, ca2 and chatting contribution!!

speccylad (from Manchester UK) chatting contribution!!

morrison_nm (@twitch) chatting contribution!!

velusip watching/viewership contribution!!

leoniietv (mrsteinberg…) chatting contribution!!

diegobrando_linux (Bl4ck_gookoo) chatting contribution!!

Beuca streaming SpeedRun contribution!!

earend from twitch  from Canada streaming contribution!!

evanmmo (@twitch) streaming contribution!!

andymilonakis from twitch Just Chatting streaming contribution!!

therealshookon3 (from twitch / from Australia) streaming contribution!!

jhovgaard streaming contribution!!

Klay4_ chatting contribution!!

HonestDanGames streaming contribution!!

NorthSeaHero streaming contribution!!

Trainwreckstv and friends streaming contribution!!

togglebit, GexYT and GoPirateSoftware streaming contribution!!

taiyoinoue, RetroMMO, OfficialAndyPyro and david_joffe streaming contribution!!

Tjienta streaming contribution!!

Primeagen streaming contribution!!

Jaxstyle and friends streaming contribution!!

EduardRFS streaming contribution!!

Melchizedek6809 and btcfly streaming contribution!!

Llama0x0 and sov_l chatting contribution!!

Tectuma (Kevin, Goth, Tim and Chronos(Math)) contribution!!

TaleLearnCode streaming contribution!!

Carol, Lucinho and Lucio contribution and visit contribution!!

hvalen_hvalborg112 streaming contribution!!

klintcsgo (CS GO: Counter-Strike Global Offensive) streaming contribution!!

KASPERPURE (Super Mario 64) streaming contribution!!

SomewhatAccurate C++ streaming contribution!!

Listening to Bryan Adams, Westlife, Shayne Ward, MLTR, Backstreet Boys, Boyzone - Best Love Songs Ever by Relax Song at YouTube!!

-- hi5 contribution...!!

Physical based rendering with Image based lightning

running Windows 10 Pro remotely from bilbo machine running Windows 10 Pro!!
canada2.camilothomas.com dedicated server by OVH.com at Beauharnois (BHS7) Canada
Intel Xeon E-2288G 8c/16t 3.7 GHz RAM 32 GB ECC 2666Hz 2x960GB NVMe SSD

+++ gpu_directx12/block.cpp
@@ -45,7 +45,7 @@ namespace gpu_directx12
 
       }
 
-      int iBufferSize = this->size(false);
+      int iBufferSize = (int) this->size(false);
 
       ::cast < renderer > prenderer = m_pgpucontext->m_pgpurenderer;
